+++ src/retail_outlet/index.js
@@ -1,0 +1,3 @@
+const RetailOutletService = require('./retail_outlet');
+
+module.exports = { RetailOutletService };