+++ src/xendit.js
@@ -6,6 +6,7 @@ const { PayoutService } = require('./payout');
 const { RecurringPayment } = require('./recurring');
 const { EWalletService } = require('./ewallet');
 const { BalanceServices } = require('./balance');
+const { RetailOutletService } = require('./retail_outlet');
 const Errors = require('./errors');
 
 function Xendit(options) {
@@ -31,6 +32,9 @@ function Xendit(options) {
   );
   this.EWallet = EWalletService._constructorWithInjectedXenditOpts(this.opts);
   this.Balance = BalanceServices._constructorWithInjectedXenditOpts(this.opts);
+  this.RetailOutlet = RetailOutletService._constructorWithInjectedXenditOpts(
+    this.opts,
+  );
 }
 
 Xendit.Errors = Errors;